requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.10.4
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
deprecation==2.1.0
Flask>=3.1.2
gotrue==2.9.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx>=0.26,<0.29
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
postgrest==0.17.0
psutil==5.9.8
psycopg2-binary>=2.9.9
pydantic>=2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-telegram-bot>=20.7
pytz==2025.2
realtime==2.7.0
six==1.17.0
sniffio==1.3.1
storage3==0.8.0
StrEnum==0.4.15
supabase>=2.8.1
supabase_auth>=2.12.3
supabase_functions==0.10.1
supafunc==0.6.0
telegram==0.0.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
websockets==15.0.1
Werkzeug==3.1.3
fastapi
psycopg2-binary
pydantic
python-jose
python-multipart
uvicorn
aiohttp
fastapi
uvicorn
aiohttp
httpx
